# Replace heuristic score multiplier map with logical operations on the coordinates