Does not affect performance significantly but makes following the logic of how heuristic piece score multipliers work much easier.

Benchmarks before:

```
[2022-06-28 20:56:51.843] [info] First move at depth 1 took 0ms and calculated a score of 0
[2022-06-28 20:56:51.932] [info] First move at depth 2 took 87ms and calculated a score of -34
[2022-06-28 20:57:14.072] [info] First move at depth 3 took 22139ms and calculated a score of 0
```

After:
```
[2022-06-28 20:59:15.491] [info] First move at depth 1 took 0ms and calculated a score of 0
[2022-06-28 20:59:15.580] [info] First move at depth 2 took 87ms and calculated a score of -34
[2022-06-28 20:59:36.034] [info] First move at depth 3 took 20453ms and calculated a score of 0
```